+++ src/main/java/absolutelyaya/ultracraft/entity/projectile/ThrownCoinEntity.java
@@ -281,7 +281,7 @@ public class ThrownCoinEntity extends ThrownItemEntity implements ProjectileEnti
 						ServerHitscanHandler.sendPacket((ServerWorld) getWorld(), getPos(), closest.getEyePos(), hitscanType);
 						closest.damage(DamageSources.get(getWorld(), DamageSources.CHARGEBACK, getOwner(), chargebackCauser), chargebackCauser == closest ? Float.MAX_VALUE : damage);
 						ExplosionHandler.explosion(getOwner(), getWorld(), closest.getPos(),
-								DamageSources.get(getWorld(), DamageTypes.EXPLOSION, this, getOwner()), 10, 0f, 5.5f, true);
+								DamageSources.get(getWorld(), DamageSources.EXPLOSION, this, getOwner()), 10, 0f, 5.5f, true);
 						Ultracraft.freeze((ServerWorld) getWorld(), 5);
 						coins.forEach(Entity::kill); //necessary because otherwise *two* final chargeback attacks occur. Don't ask why, I have no idea
 						return true;